Week 3 Workshop Assignment

+++ routes/campsiteRouter.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const Campsite = require("../models/campsite");
 const authenticate = require("../authenticate");
+const { verify } = require("jsonwebtoken");
 
 const campsiteRouter = express.Router();
 
@@ -16,29 +17,45 @@ campsiteRouter
 			})
 			.catch((err) => next(err));
 	})
-	.post(authenticate.verifyUser, (req, res, next) => {
-		Campsite.create(req.body)
-			.then((campsite) => {
-				console.log("Campsite Created ", campsite);
-				res.statusCode = 200;
-				res.setHeader("Content-Type", "application/json");
-				res.json(campsite);
-			})
-			.catch((err) => next(err));
-	})
+	.post(
+		[authenticate.verifyUser, authenticate.verifyAdmin],
+		(req, res, next) => {
+			if (req.user.admin) {
+				Campsite.create(req.body)
+					.then((campsite) => {
+						console.log("Campsite Created ", campsite);
+						res.statusCode = 200;
+						res.setHeader("Content-Type", "application/json");
+						res.json(campsite);
+					})
+					.catch((err) => next(err));
+			} else {
+				res.statusCode = 403;
+				res.end("Access denied.");
+			}
+		}
+	)
 	.put(authenticate.verifyUser, (req, res) => {
 		res.statusCode = 403;
 		res.end("PUT operation not supported on /campsites");
 	})
-	.delete(authenticate.verifyUser, (req, res, next) => {
-		Campsite.deleteMany()
-			.then((response) => {
-				res.statusCode = 200;
-				res.setHeader("Content-Type", "application/json");
-				res.json(response);
-			})
-			.catch((err) => next(err));
-	});
+	.delete(
+		[authenticate.verifyUser, authenticate.verifyAdmin],
+		(req, res, next) => {
+			if (req.user.admin) {
+				Campsite.deleteMany()
+					.then((response) => {
+						res.statusCode = 200;
+						res.setHeader("Content-Type", "application/json");
+						res.json(response);
+					})
+					.catch((err) => next(err));
+			} else {
+				res.statusCode = 403;
+				res.end("Access denied.");
+			}
+		}
+	);
 
 campsiteRouter
 	.route("/:campsiteId")
@@ -58,30 +75,46 @@ campsiteRouter
 			`POST operation not supported on /campsites/${req.params.campsiteId}`
 		);
 	})
-	.put(authenticate.verifyUser, (req, res, next) => {
-		Campsite.findByIdAndUpdate(
-			req.params.campsiteId,
-			{
-				$set: req.body,
-			},
-			{ new: true }
-		)
-			.then((campsite) => {
-				res.statusCode = 200;
-				res.setHeader("Content-Type", "application/json");
-				res.json(campsite);
-			})
-			.catch((err) => next(err));
-	})
-	.delete(authenticate.verifyUser, (req, res, next) => {
-		Campsite.findByIdAndDelete(req.params.campsiteId)
-			.then((response) => {
-				res.statusCode = 200;
-				res.setHeader("Content-Type", "application/json");
-				res.json(response);
-			})
-			.catch((err) => next(err));
-	});
+	.put(
+		[authenticate.verifyUser, authenticate.verifyAdmin],
+		(req, res, next) => {
+			if (req.user.admin) {
+				Campsite.findByIdAndUpdate(
+					req.params.campsiteId,
+					{
+						$set: req.body,
+					},
+					{ new: true }
+				)
+					.then((campsite) => {
+						res.statusCode = 200;
+						res.setHeader("Content-Type", "application/json");
+						res.json(campsite);
+					})
+					.catch((err) => next(err));
+			} else {
+				res.statusCode = 403;
+				res.end("Access denied.");
+			}
+		}
+	)
+	.delete(
+		[authenticate.verifyUser, authenticate.verifyAdmin],
+		(req, res, next) => {
+			if (req.user.admin) {
+				Campsite.findByIdAndDelete(req.params.campsiteId)
+					.then((response) => {
+						res.statusCode = 200;
+						res.setHeader("Content-Type", "application/json");
+						res.json(response);
+					})
+					.catch((err) => next(err));
+			} else {
+				res.statusCode = 403;
+				res.end("Access denied.");
+			}
+		}
+	);
 
 campsiteRouter
 	.route("/:campsiteId/comments")
@@ -129,29 +162,37 @@ campsiteRouter
 			`PUT operation not supported on /campsites/${req.params.campsiteId}/comments`
 		);
 	})
-	.delete(authenticate.verifyUser, (req, res, next) => {
-		Campsite.findById(req.params.campsiteId)
-			.then((campsite) => {
-				if (campsite) {
-					for (let i = campsite.comments.length - 1; i >= 0; i--) {
-						campsite.comments.id(campsite.comments[i]._id).remove();
-					}
-					campsite
-						.save()
-						.then((campsite) => {
-							res.statusCode = 200;
-							res.setHeader("Content-Type", "application/json");
-							res.json(campsite);
-						})
-						.catch((err) => next(err));
-				} else {
-					err = new Error(`Campsite ${req.params.campsiteId} not found`);
-					err.status = 404;
-					return next(err);
-				}
-			})
-			.catch((err) => next(err));
-	});
+	.delete(
+		[authenticate.verifyUser, authenticate.verifyAdmin],
+		(req, res, next) => {
+			if (req.user.admin) {
+				Campsite.findById(req.params.campsiteId)
+					.then((campsite) => {
+						if (campsite) {
+							for (let i = campsite.comments.length - 1; i >= 0; i--) {
+								campsite.comments.id(campsite.comments[i]._id).remove();
+							}
+							campsite
+								.save()
+								.then((campsite) => {
+									res.statusCode = 200;
+									res.setHeader("Content-Type", "application/json");
+									res.json(campsite);
+								})
+								.catch((err) => next(err));
+						} else {
+							err = new Error(`Campsite ${req.params.campsiteId} not found`);
+							err.status = 404;
+							return next(err);
+						}
+					})
+					.catch((err) => next(err));
+			} else {
+				res.statusCode = 403;
+				res.end("Access denied.");
+			}
+		}
+	);
 
 campsiteRouter
 	.route("/:campsiteId/comments/:commentId")
@@ -182,59 +223,70 @@ campsiteRouter
 		);
 	})
 	.put(authenticate.verifyUser, (req, res, next) => {
-		Campsite.findById(req.params.campsiteId)
-			.then((campsite) => {
-				if (campsite && campsite.comments.id(req.params.commentId)) {
-					if (req.body.rating) {
-						campsite.comments.id(req.params.commentId).rating = req.body.rating;
+		if (req.user._id) {
+			Campsite.findById(req.params.campsiteId)
+				.then((campsite) => {
+					if (campsite && campsite.comments.id(req.params.commentId)) {
+						if (req.body.rating) {
+							campsite.comments.id(req.params.commentId).rating =
+								req.body.rating;
+						}
+						if (req.body.text) {
+							campsite.comments.id(req.params.commentId).text = req.body.text;
+						}
+						campsite
+							.save()
+							.then((campsite) => {
+								res.statusCode = 200;
+								res.setHeader("Content-Type", "application/json");
+								res.json(campsite);
+							})
+							.catch((err) => next(err));
+					} else if (!campsite) {
+						err = new Error(`Campsite ${req.params.campsiteId} not found`);
+						err.status = 404;
+						return next(err);
+					} else {
+						err = new Error(`Comment ${req.params.commentId} not found`);
+						err.status = 404;
+						return next(err);
 					}
-					if (req.body.text) {
-						campsite.comments.id(req.params.commentId).text = req.body.text;
-					}
-					campsite
-						.save()
-						.then((campsite) => {
-							res.statusCode = 200;
-							res.setHeader("Content-Type", "application/json");
-							res.json(campsite);
-						})
-						.catch((err) => next(err));
-				} else if (!campsite) {
-					err = new Error(`Campsite ${req.params.campsiteId} not found`);
-					err.status = 404;
-					return next(err);
-				} else {
-					err = new Error(`Comment ${req.params.commentId} not found`);
-					err.status = 404;
-					return next(err);
-				}
-			})
-			.catch((err) => next(err));
+				})
+				.catch((err) => next(err));
+		} else {
+			res.statusCode = 403;
+			res.end("Access denied.");
+		}
 	})
 	.delete(authenticate.verifyUser, (req, res, next) => {
-		Campsite.findById(req.params.campsiteId)
-			.then((campsite) => {
-				if (campsite && campsite.comments.id(req.params.commentId)) {
-					campsite.comments.id(req.params.commentId).remove();
-					campsite
-						.save()
-						.then((campsite) => {
-							res.statusCode = 200;
-							res.setHeader("Content-Type", "application/json");
-							res.json(campsite);
-						})
-						.catch((err) => next(err));
-				} else if (!campsite) {
-					err = new Error(`Campsite ${req.params.campsiteId} not found`);
-					err.status = 404;
-					return next(err);
-				} else {
-					err = new Error(`Comment ${req.params.commentId} not found`);
-					err.status = 404;
-					return next(err);
-				}
-			})
-			.catch((err) => next(err));
+		if (req.user._id) {
+			Campsite.findById(req.params.campsiteId)
+				.then((campsite) => {
+					if (campsite && campsite.comments.id(req.params.commentId)) {
+						campsite.comments.id(req.params.commentId).remove();
+						campsite
+							.save()
+							.then((campsite) => {
+								res.statusCode = 200;
+								res.setHeader("Content-Type", "application/json");
+								res.json(campsite);
+							})
+							.catch((err) => next(err));
+					} else if (!campsite) {
+						err = new Error(`Campsite ${req.params.campsiteId} not found`);
+						err.status = 404;
+						return next(err);
+					} else {
+						err = new Error(`Comment ${req.params.commentId} not found`);
+						err.status = 404;
+						return next(err);
+					}
+				})
+				.catch((err) => next(err));
+		} else {
+			res.statusCode = 403;
+			res.end("Access denied.");
+		}
 	});
 
 module.exports = campsiteRouter;
