+++ routes/partnerRouter.js
@@ -15,29 +15,45 @@ partnerRouter
 			})
 			.catch((err) => next(err));
 	})
-	.post(authenticate.verifyUser, (req, res, next) => {
-		Partner.create(req.body)
-			.then((partner) => {
-				console.log("Partner Created", partner);
-				res.statusCode = 200;
-				res.setHeader("Content-Type", "application/json");
-				res.json(partner);
-			})
-			.catch((err) => next(err));
-	})
+	.post(
+		[authenticate.verifyUser, authenticate.verifyAdmin],
+		(req, res, next) => {
+			if (req.user.admin) {
+				Partner.create(req.body)
+					.then((partner) => {
+						console.log("Partner Created", partner);
+						res.statusCode = 200;
+						res.setHeader("Content-Type", "application/json");
+						res.json(partner);
+					})
+					.catch((err) => next(err));
+			} else {
+				res.statusCode = 403;
+				res.end("Access denied.");
+			}
+		}
+	)
 	.put(authenticate.verifyUser, (req, res) => {
 		res.statusCode = 403;
 		res.end("PUT operation not supported on /partners");
 	})
-	.delete(authenticate.verifyUser, (req, res, next) => {
-		Partner.deleteMany()
-			.then((response) => {
-				res.statusCode = 200;
-				res.setHeader("Content-Type", "application/json");
-				res.json(response);
-			})
-			.catch((err) => next(err));
-	});
+	.delete(
+		[authenticate.verifyUser, authenticate.verifyAdmin],
+		(req, res, next) => {
+			if (req.user.admin) {
+				Partner.deleteMany()
+					.then((response) => {
+						res.statusCode = 200;
+						res.setHeader("Content-Type", "application/json");
+						res.json(response);
+					})
+					.catch((err) => next(err));
+			} else {
+				res.statusCode = 403;
+				res.end("Access denied.");
+			}
+		}
+	);
 
 partnerRouter
 	.route("/:partnerId")
@@ -56,29 +72,45 @@ partnerRouter
 			`POST operation not supported on /partners/${req.params.partnerId}`
 		);
 	})
-	.put(authenticate.verifyUser, (req, res, next) => {
-		Partner.findByIdAndUpdate(
-			req.params.partnerId,
-			{
-				$set: req.body,
-			},
-			{ new: true }
-		)
-			.then((partner) => {
-				res.statusCode = 200;
-				res.setHeader("Content-Type", "application/json");
-				res.json(partner);
-			})
-			.catch((err) => next(err));
-	})
-	.delete(authenticate.verifyUser, (req, res, next) => {
-		Partner.findByIdAndDelete(req.params.partnerId)
-			.then((response) => {
-				res.statusCode = 200;
-				res.setHeader("Content-Type", "application/json");
-				res.json(response);
-			})
-			.catch((err) => next(err));
-	});
+	.put(
+		[authenticate.verifyUser, authenticate.verifyAdmin],
+		(req, res, next) => {
+			if (req.user.admin) {
+				Partner.findByIdAndUpdate(
+					req.params.partnerId,
+					{
+						$set: req.body,
+					},
+					{ new: true }
+				)
+					.then((partner) => {
+						res.statusCode = 200;
+						res.setHeader("Content-Type", "application/json");
+						res.json(partner);
+					})
+					.catch((err) => next(err));
+			} else {
+				res.statusCode = 403;
+				res.end("Access denied.");
+			}
+		}
+	)
+	.delete(
+		[authenticate.verifyUser, authenticate.verifyAdmin],
+		(req, res, next) => {
+			if (req.user.admin) {
+				Partner.findByIdAndDelete(req.params.partnerId)
+					.then((response) => {
+						res.statusCode = 200;
+						res.setHeader("Content-Type", "application/json");
+						res.json(response);
+					})
+					.catch((err) => next(err));
+			} else {
+				res.statusCode = 403;
+				res.end("Access denied.");
+			}
+		}
+	);
 
 module.exports = partnerRouter;
